+++ jassrealtime/search/multicorpus/documents_by_text.py
@@ -1,14 +1,13 @@
 from elasticsearch_dsl import Search, Q
 
+from .DocumentsBy import DocumentsBy
 from ...search.multicorpus.multi_corpus import MultiCorpus
-from ...core.language_manager import LanguageManager
-from ...core.settings_utils import get_language_manager
 from ...core.esutils import get_es_conn
 from ...search.document import map_search_hit
 from ...security.base_authorization import BaseAuthorization
 
 
-class DocumentsByText:
+class DocumentsByText(DocumentsBy):
     def __init__(self, env_id: str, authorization: BaseAuthorization):
         self.env_id = env_id
         self.authorization = authorization
@@ -26,29 +25,6 @@ class DocumentsByText:
 
     def queries_indices(self, queries: list) -> list:
         return [self.query_index(query) for query in queries]
-
-    @staticmethod
-    def transform(language_manager: LanguageManager, query: dict) -> Q:
-        text_field_name = "text"
-        if query["search_mode"] == "language":
-            text_field_name = text_field_name + "." + language_manager.get_es_analyser(query["language"])
-
-        return Q({"match": {text_field_name: query["text"]}})
-
-    def group_and_transform_queries_by_operator(self, queries: list) -> dict:
-        language_manager = get_language_manager()
-        operators = ["must", "must_not", "should"]
-        grouped_queries = {}
-        for operator in operators:
-            grouped_queries[operator] = []
-
-        for query in queries:
-            for operator in operators:
-                if operator == query["operator"]:
-                    grouped_queries[operator].append(self.transform(language_manager, query))
-                    continue
-
-        return grouped_queries
 
     def documents_by_text(self, queries: list, from_index: int, size: int) -> tuple:
         """
